CHG: move ICA tests to separate file.
